backend/requirements.txt
Flask==2.3.3
google-generativeai>=0.8.0
chromadb
nltk
Werkzeug
flask-cors
